AWS/watchdog_cloud_server/requirements.txt
Django==2.2
djongo==1.3.6
dnspython==2.3.0
pymongo==4.4.1
pytz==2023.3
sqlparse==0.2.4
mongoengine
requests==2.25.1
paho-mqtt==1.5.1
